+++ src/KindsToFlavors.jsx
@@ -1,7 +1,16 @@
 import { Link } from "@material-ui/icons";
 import { parse } from "query-string";
 import React from "react";
-import { Create, DisabledInput, Edit, ReferenceInput, SelectInput, SimpleForm, required } from "react-admin";
+import {
+    Create,
+    DisabledInput,
+    Edit,
+    ReferenceInput,
+    SelectInput,
+    SimpleForm,
+    required,
+    AutocompleteInput
+} from "react-admin";
 
 export const KindsToFlavorsIcon = Link;
 
@@ -15,8 +24,8 @@ export const KindsToFlavorsEdit = props => (
     <Edit title={<KindsToFlavorsTitle />} {...props}>
         <SimpleForm redirect={redirect}>
             <DisabledInput source="id" />
-            <ReferenceInput source="flavor_id" reference="flavors" perPage={100} validate={required()}>
-                <SelectInput optionText="name" />
+            <ReferenceInput source="flavor_id" reference="flavors" perPage={50} sort={{"field": "name"}} validate={required()}>
+                <AutocompleteInput optionText="name" translateChoice={false} />
             </ReferenceInput>
         </SimpleForm>
     </Edit>
@@ -28,11 +37,11 @@ export const KindsToFlavorsCreate = props => {
     return (
         <Create title="Add a Flavor" {...props}>
             <SimpleForm redirect={redirect} defaultValue={{ kind_id }}>
-                <ReferenceInput source="kind_id" reference="kinds" perPage={100} validate={required()}>
+                <ReferenceInput source="kind_id" reference="kinds" perPage={50} sort={{"field": "name"}} validate={required()}>
                     <SelectInput optionText="name" />
                 </ReferenceInput>
-                <ReferenceInput source="flavor_id" reference="flavors" perPage={100} validate={required()}>
-                    <SelectInput optionText="name" />
+                <ReferenceInput source="flavor_id" reference="flavors" sort={{"field": "name"}} validate={required()}>
+                    <AutocompleteInput optionText="name" translateChoice={false} />
                 </ReferenceInput>
             </SimpleForm>
         </Create>